Refactored a lot of things, particularly concerning the structs and enums: when possible I define structs in the enums variant, this makes things way simpler and easy to read. Also gave up and making an hyper compatible and shared structures in the bridge -> we'll see later how to handle that, for now I design a nice API

+++ dmlogic/src/moms/mod.rs
@@ -1,13 +1,11 @@
 use crate::protocol::operations::Operation;
-use crate::bridge::ffi::TaskSend;
-use crate::motherboard::MotherboardOrMOMEvent;
-
+use crate::motherboard::{MotherboardOrMOMEvent, TaskSend};
 
 pub enum MOMEvent {
     OperationReceived(Operation),
     NodeConnected(),
     // Contains the number of connected clients
-    ClusterConnected(u32),
+    ClusterConnected(u64),
 }
 
 pub trait MOM {
